requirements.txt
# Django e REST Framework
Django>=5.1,<7.0
djangorestframework>=3.15
django-cors-headers>=4.6

# LangChain — orquestração de IA
langchain>=1.2
langchain-openai>=1.1
langchain-community>=0.3

# LangSmith — monitoramento de chamadas LLM (opcional)
langsmith>=0.7

# OpenAI SDK (versão gerenciada pelo langchain-openai)
openai>=2.0,<3.0

# Variáveis de ambiente
python-dotenv>=1.0
